Refactor the "place" step to use reusable PageBase component
